+++ .env
@@ -2,7 +2,10 @@ NEXT_PUBLIC_KAKAO_API_KEY=8f4ed639e0b0d07de0e3ce0c3365f726
 NEXT_PUBLIC_KAKAO_REST_API_KEY=265b1c201bbb3dfc32f986bccfbf8f23E
 TWITTER_CLIENT_ID=aXBlT0w0Zl9CYUpQZUxaemtmRkY6MTpjaQ
 TWITTER_CLIENT_SECRET=h6iEg2jR1_trC0TelEb50EUVEl7lAjz2HE9tZK1fD2AYryuBYY
+KAKAO_CLIENT_ID=8f4ed639e0b0d07de0e3ce0c3365f726
+KAKAO_CLIENT_SECRET=dsfsdfdsfdsfevdffrdfefefdfdfdf27362hehcydgsjwj7
 NEXTAUTH_SECRET=qerkasjhfkauiqeywrhkasjhfksajfh
 SECRET_KEY=dkasdhfjahsdfjhskdfhaksdhfakjs
 REGISTRY_HOST=gk1vk
 NEXT_PUBLIC_API_HOST=http://141.147.166.132:3100
+NEXTAUTH_URL=http://local.named.co
