--- src/api/auth/[...nextauth].ts
@@ -1,12 +1,0 @@
-import NextAuth from 'next-auth';
-import TwitterProvider from 'next-auth/providers/twitter';
-
-export default NextAuth({
-  providers: [
-    TwitterProvider({
-      clientId: process.env.TWITTER_CLIENT_ID || '',
-      clientSecret: process.env.TWITTER_CLIENT_SECRET || '',
-      version: '2.0',
-    }),
-  ],
-});
