+++ src/components/login/SubmitLoginButton.tsx
@@ -1,8 +1,7 @@
 import { Button } from '@components/common';
-import { useUserLoginErrorStore } from '@zustand/userLoginStore';
+import { useUserLoginErrorStore } from '@zustand/usersLoginStore';
 import { useRouter } from 'next/router';
 import { signInEmail } from 'pages/api/login/login';
-import { useEffect } from 'react';
 import { loginInputValidation } from 'validation/loginValidation';
 
 interface SubmitLoginButtonProps {
