+++ src/pages/api/auth/[...nextauth].ts
@@ -1,4 +1,5 @@
 import NextAuth from "next-auth";
+import KakaoProvider from "next-auth/providers/kakao";
 import TwitterProvider from "next-auth/providers/twitter";
 
 export default NextAuth({
@@ -8,5 +9,9 @@ export default NextAuth({
       clientSecret: process.env.TWITTER_CLIENT_SECRET || "",
       version: "2.0",
     }),
+    KakaoProvider({
+      clientId: process.env.KAKAO_CLIENT_ID || "",
+      clientSecret: process.env.KAKAO_CLIENT_SECRET || ""
+    })
   ],
 });
